+++ apps/decodex/src/agent/tracker_tool_bridge/tests/mutation/continuation/completion_conflicts.rs
@@ -1,0 +1,60 @@
+use tempfile::TempDir;
+
+use crate::agent::tracker_tool_bridge::{
+	DynamicToolHandler, ISSUE_LABEL_ADD_TOOL_NAME, ISSUE_REVIEW_HANDOFF_TOOL_NAME,
+	PullRequestDetails, TrackerToolBridge,
+	tests::{self, FakeLocalRepoInspector, FakePullRequestInspector},
+};
+
+#[test]
+fn completion_disposition_rejects_conflicting_review_handoff_and_manual_attention() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let issue = tests::sample_issue();
+	let tracker = tests::tracker_with_current_issue_snapshot(&issue);
+	let workflow = tests::sample_workflow();
+	let inspector = FakePullRequestInspector::new(vec![Ok(PullRequestDetails {
+		head_ref_name: String::from("x/decodex-pub-618"),
+		head_ref_oid: String::from("08a20f7dfb9526e7421a5f095b1c6adec84e52d6"),
+		head_repository_name: String::from("decodex"),
+		head_repository_owner: String::from("hack-ink"),
+		is_draft: false,
+		state: String::from("OPEN"),
+		base_ref_name: String::from("main"),
+		url: String::from("https://github.com/hack-ink/decodex/pull/48"),
+	})]);
+	let local_repo_inspector = FakeLocalRepoInspector::new(vec![Ok(tests::sample_local_repo())]);
+	let review_context = tests::sample_review_context_in(temp_dir.path());
+	let bridge = TrackerToolBridge::with_review_handoff_for_test(
+		&tracker,
+		&issue,
+		&workflow,
+		review_context.clone(),
+		&inspector,
+		&local_repo_inspector,
+	);
+
+	tests::write_clean_review_checkpoint(&bridge, &issue, &review_context);
+
+	let review_response = DynamicToolHandler::handle_call(
+		&bridge,
+		ISSUE_REVIEW_HANDOFF_TOOL_NAME,
+		serde_json::json!({
+			"pr_url": "https://github.com/hack-ink/decodex/pull/48",
+			"summary": "Ready for review."
+		}),
+	);
+	let label_response = DynamicToolHandler::handle_call(
+		&bridge,
+		ISSUE_LABEL_ADD_TOOL_NAME,
+		serde_json::json!({ "label": "decodex:needs-attention" }),
+	);
+
+	assert!(review_response.success);
+	assert!(label_response.success);
+
+	let error = bridge
+		.completion_disposition()
+		.expect_err("conflicting completion signals should be rejected");
+
+	assert!(error.to_string().contains("Use exactly one final tracker exit path."));
+}
